agent-service/requirements.txt
# Agent Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Templates & Static Files
jinja2==3.1.3
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# LLM Integration
anthropic==0.18.1
openai==1.12.0

# Cryptography (for TAP signing)
cryptography==42.0.0
PyJWT==2.8.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# WebSocket for real-time chat
websockets==12.0
